+++ packages/client/hmi-client/tests/unit/utils/operator-bitmask.spec.ts
@@ -1,4 +1,11 @@
-import { addHover, removeHover, isHover, addDrag, removeDrag, isDrag } from '@/utils/bitmask';
+import {
+	addHover,
+	removeHover,
+	isHover,
+	addDrag,
+	removeDrag,
+	isDrag
+} from '@/services/operator-bitmask';
 import { describe, expect, it } from 'vitest';
 
 let interactionStatus = 0;
